requirements.txt
pyotp
playwright
beautifulsoup4
python-dotenv